Move 'attribute' to private folder
