+++ addon/index.js
@@ -10,7 +10,7 @@ import { hasClass }    from './-private/properties/has-class';        export { h
 import { isHidden }    from './-private/properties/is-hidden';        export { isHidden };
 import { isVisible }   from './-private/properties/is-visible';       export { isVisible };
 import { notHasClass } from './-private/properties/not-has-class';    export { notHasClass };
-import { attribute }   from './queries/attribute';        export { attribute };
+import { attribute }   from './-private/properties/attribute';        export { attribute };
 import { count }       from './queries/count';            export { count };
 import { is }          from './queries/is';               export { is };
 import { property }    from './queries/property';         export { property };
